add many many things

+++ src/app/features/favorite/services/favorite.service.ts
@@ -1,3 +1,4 @@
+import { ToastrService } from 'ngx-toastr';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
@@ -13,41 +14,76 @@ export class FavoriteService {
   controllerUrl: string;
 
   constructor(private accountService: AccountService,
-              private http: HttpClient) {
+              private http: HttpClient,
+              private toastr:ToastrService) {
     this.controllerUrl = environment.apiUrl + "favorite/";
    }
 
   addFavoritePlace(id:number):Observable<any>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      this.toastr.error("You are not authorized!!!");
+      return new Observable<1>;
+    }
+
     return this.http.post(this.controllerUrl+"add-place/"+id,null,this.accountService.getHttpOptions());
   }
 
   addFavoriteEvent(id:number):Observable<any>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      this.toastr.error("You are not authorized!!!");
+      return new Observable<1>;
+    }
+
     return this.http.post(this.controllerUrl+"add-event/"+id,null,this.accountService.getHttpOptions());
   }
 
 
   removeFavoritePlace(id:number):Observable<any>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      this.toastr.error("You are not authorized!!!");
+      return new Observable<1>;
+    }
+
     return this.http.delete(this.controllerUrl+"remove-place/"+id,this.accountService.getHttpOptions());
   }
 
   removeFavoriteEvent(id:number):Observable<any>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      this.toastr.error("You are not authorized!!!");
+      return new Observable<1>;
+    }
+
     return this.http.delete(this.controllerUrl+"remove-event/"+id,this.accountService.getHttpOptions());
   }
 
 
   getFavoritePlaces():Observable<IPlace[]>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      return new Observable<[]>;
+    }
+
     const httpOptions = {
         headers: new HttpHeaders({
-            Authorization: 'Bearer '+ this.accountService.getToken()
+            Authorization: 'Bearer '+ token
         })
     };
     return this.http.get<IPlace[]>(this.controllerUrl+"get-places",httpOptions);
   }
 
   getFavoriteEvents():Observable<IEvent[]>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      return new Observable<[]>;
+    }
+
     const httpOptions = {
         headers: new HttpHeaders({
-            Authorization: 'Bearer '+ this.accountService.getToken()
+            Authorization: 'Bearer '+ token
         })
     };
     return this.http.get<IEvent[]>(this.controllerUrl+"get-events",httpOptions);
@@ -55,18 +91,29 @@ export class FavoriteService {
 
 
   isFavoritePlace(id:number):Observable<boolean>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      return new Observable<false>;
+    }
+
+
     const httpOptions = {
         headers: new HttpHeaders({
-            Authorization: 'Bearer '+ this.accountService.getToken()
+            Authorization: 'Bearer '+ token
         })
     };
     return this.http.get<boolean>(this.controllerUrl+"is-place/"+id,httpOptions);
   }
 
   isFavoriteEvent(id:number):Observable<boolean>{
+    let token = this.accountService.getToken();
+    if(token==null){
+      return new Observable<false>;
+    }
+
     const httpOptions = {
         headers: new HttpHeaders({
-            Authorization: 'Bearer '+ this.accountService.getToken()
+            Authorization: 'Bearer '+ token
         })
     };
     return this.http.get<boolean>(this.controllerUrl+"is-event/"+id,httpOptions);
